+++ tests/Domain/Service/PaymentCalculatorTest.php
@@ -155,11 +155,35 @@ final class PaymentCalculatorTest extends TestCase
     {
         $this->statusForAll(PaymentStatus::HUMAN_SCORE_TOO_LOW, ['human_score' => 0]);
         $this->statusForAll(PaymentStatus::HUMAN_SCORE_TOO_LOW, ['human_score' => 0.3]);
-        $this->statusForAll(PaymentStatus::HUMAN_SCORE_TOO_LOW, ['human_score' => 0.499]);
+        $this->statusForAll(PaymentStatus::HUMAN_SCORE_TOO_LOW, ['human_score' => 0.399]);
         $this->statusForAll(PaymentStatus::ACCEPTED, ['human_score' => 0.5]);
         $this->statusForAll(PaymentStatus::ACCEPTED, ['human_score' => 0.501]);
         $this->statusForAll(PaymentStatus::ACCEPTED, ['human_score' => 0.7]);
         $this->statusForAll(PaymentStatus::ACCEPTED, ['human_score' => 1]);
+
+        $campaigns = new CampaignCollection(self::campaign([], [self::banner()], [self::conversion()]));
+        $payment = $this->single($campaigns, self::viewEvent(['human_score' => 0.499]));
+        $this->assertEquals(PaymentStatus::HUMAN_SCORE_TOO_LOW, $payment['status']);
+
+        $campaigns = new CampaignCollection(self::campaign([], [self::banner()], [self::conversion()]));
+        $payment = $this->single($campaigns, self::viewEvent(['human_score' => 0.4]));
+        $this->assertEquals(PaymentStatus::HUMAN_SCORE_TOO_LOW, $payment['status']);
+
+        $campaigns = new CampaignCollection(self::campaign([], [self::banner()], [self::conversion()]));
+        $payment = $this->single($campaigns, self::clickEvent(['human_score' => 0.499]));
+        $this->assertEquals(PaymentStatus::HUMAN_SCORE_TOO_LOW, $payment['status']);
+
+        $campaigns = new CampaignCollection(self::campaign([], [self::banner()], [self::conversion()]));
+        $payment = $this->single($campaigns, self::clickEvent(['human_score' => 0.4]));
+        $this->assertEquals(PaymentStatus::HUMAN_SCORE_TOO_LOW, $payment['status']);
+
+        $campaigns = new CampaignCollection(self::campaign([], [self::banner()], [self::conversion()]));
+        $payment = $this->single($campaigns, self::conversionEvent(['human_score' => 0.499]));
+        $this->assertEquals(PaymentStatus::ACCEPTED, $payment['status']);
+
+        $campaigns = new CampaignCollection(self::campaign([], [self::banner()], [self::conversion()]));
+        $payment = $this->single($campaigns, self::conversionEvent(['human_score' => 0.4]));
+        $this->assertEquals(PaymentStatus::ACCEPTED, $payment['status']);
     }
 
     public function testHumanScoreThreshold(): void
@@ -168,14 +192,22 @@ final class PaymentCalculatorTest extends TestCase
 
         $payment = $this->single($campaigns, self::viewEvent(['human_score' => 0.5]));
         $this->assertEquals(PaymentStatus::ACCEPTED, $payment['status']);
+        $payment = $this->single($campaigns, self::conversionEvent(['human_score' => 0.4]));
+        $this->assertEquals(PaymentStatus::ACCEPTED, $payment['status']);
 
         $payment = $this->single($campaigns, self::viewEvent(['human_score' => 0.5]), ['humanScoreThreshold' => 0.55]);
+        $this->assertEquals(PaymentStatus::HUMAN_SCORE_TOO_LOW, $payment['status']);
+        $payment = $this->single($campaigns, self::conversionEvent(['human_score' => 0.5]), ['conversionHumanScoreThreshold' => 0.55]);
         $this->assertEquals(PaymentStatus::HUMAN_SCORE_TOO_LOW, $payment['status']);
 
         $payment = $this->single($campaigns, self::viewEvent(['human_score' => 0.3]), ['humanScoreThreshold' => '0.5']);
         $this->assertEquals(PaymentStatus::HUMAN_SCORE_TOO_LOW, $payment['status']);
+        $payment = $this->single($campaigns, self::conversionEvent(['human_score' => 0.3]), ['humanScoreThreshold' => '0.5']);
+        $this->assertEquals(PaymentStatus::HUMAN_SCORE_TOO_LOW, $payment['status']);
 
         $payment = $this->single($campaigns, self::viewEvent(['human_score' => 0.49]), ['humanScoreThreshold' => null]);
+        $this->assertEquals(PaymentStatus::HUMAN_SCORE_TOO_LOW, $payment['status']);
+        $payment = $this->single($campaigns, self::conversionEvent(['human_score' => 0.39]), ['humanScoreThreshold' => '0.5']);
         $this->assertEquals(PaymentStatus::HUMAN_SCORE_TOO_LOW, $payment['status']);
     }