+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.1] - 2019-12-03
 ### Changed 
 - Case time support
+- Lowering the human score threshold for conversions
 
 ## [1.1.0] - 2019-12-02
 ### Added